fix(tasks): harden completion and shortage workflow
Four defects, all found by looking at what Adım 15 actually made
reachable rather than at what it wrote.

The pool card counted failures into an Int. The database adds them up
as a 64 bit number and PLAN 6.4 leaves that total deliberately
uncapped, so three reports of the largest amount the form will take
came back to the card as a negative count of things that had gone
wrong — and a negative one is hidden rather than shown. Nothing could
write a failure until this step; now the projection is as wide as the
sum it carries.

A shortage could still be reported into a finished game. Refusing to
open the form is not the whole rule when the game can be finished
while the form stands open, so the same question is asked again where
the write happens. Making good is deliberately still allowed: a task
that owes something was never finished, so settling it reopens
nothing.

Two presses of a tick meant as one undid each other. The write comes
back long before the row carrying its result does, and in that gap the
tick is still drawn in its old state, so the second press read what it
was about to leave and asked for the opposite — leaving a task that
looked untouched while the shortage it owed had been settled into its
history for good. The tick is now held until the table has answered.

The making good form offered a step chooser, and nothing is kept about
where a piece was made good again. It took an answer and dropped it.

+++ app/src/commonMain/kotlin/dev/pnptracker/data/database/projection/PoolRows.kt
@@ -79,8 +79,18 @@ data class PoolStageRow(
 data class PoolFailureRow(
     @ColumnInfo(name = "task_id")
     val taskId: EntityId,
+    /**
+     * Held as a `Long` because that is what the database added up.
+     *
+     * `SUM` over the events is a 64 bit number and the failure total is
+     * deliberately uncapped — PLAN 6.4 lets a piece be spoiled more times than
+     * the task needs pieces — so there is nothing here to bound it by. Narrowing
+     * it would not report a number too large, it would report a different one:
+     * the low half of it, which past two thousand million reads as a negative
+     * count of things that went wrong.
+     */
     @ColumnInfo(name = "failure_total")
-    val failureTotal: Int,
+    val failureTotal: Long,
 )
 
 /** How much work each pool is holding, for the sidebar. */
